# Reduce memory inefficiencies and leaks