--- a/Source/Noesis.Javascript/JavascriptExternal.h
+++ b/Source/Noesis.Javascript/JavascriptExternal.h
@@ -31,7 +31,6 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include <v8.h>
-#include <map>
 #include <gcroot.h>
 #include "JavascriptContext.h"
 
@@ -101,9 +100,6 @@ private:
 	System::Runtime::InteropServices::GCHandle mObjectHandle;
 
 	SetParameterOptions mOptions;
-
-	// Owned by JavascriptContext.
-	gcroot<System::Collections::Generic::Dictionary<System::String ^, WrappedMethod> ^> mMethods;
 
     std::unique_ptr<Persistent<Function>> mIterator;
     static void IteratorCallback(const v8::FunctionCallbackInfo<Value>& iArgs);